* walls are correctly positioned in 3D now

+++ src/view3d.c
@@ -201,8 +201,8 @@ gboolean view3d_motion_cb(GtkWidget *widget, GdkEventMotion *event,
 		/* left button pressed */
 		if(state & GDK_SHIFT_MASK) {
 			/* shift pressed, translate view */
-			house->off_x += (x - priv->drag_x) / 10;
-			house->off_y += (y - priv->drag_y) / 10;
+			house->off_x += (gdouble)(x - priv->drag_x) / 10.0;
+			house->off_y += (gdouble)(y - priv->drag_y) / 10.0;
 		} else {
 			/* rotate view */
 			gfloat spin_quat[4];
